+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -9,12 +9,30 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+5.0.0.6
+-------
+Added
+*****
+- Added AddGameFinder extension method to create a gamefinder scene from a gamefinder configuration
+- Add parametrized quick queue implementations (static & dynamic party parameters)
+- Add support for canPlayTogether filters in the quick queue implementation.
+- Add support for setting a custom gamefinder party parameters factory function for the quick queue implementation. 
+
+Changed
+*******
+- Change AddGameFinder extension method to ConfigureGameFinder because it doesn't actually creates the gamefinder scene, just the template.
+- Correctly dispose packet in S2S RPC
+
 4.0.2
 -----
+Changed
+*******
 - Use Stormancer.Abstraction.Server 6.1.0
 
 4.0.1.1
 -------
+Changed
+*******
 - Create GameFinderService dependency at the scene level to prevent it from being instantiated in all scenes by IConfigurationChangedEventHandler notifications.
 
 4.0.0.2